requirements.txt
# requirements.txt
fastapi
uvicorn
azure-servicebus
python-dotenv
azure-search-documents==11.4.0
langchain-openai==0.2.3
langgraph==0.2.56
azure-ai-inference[opentelemetry]